MusicPlayer - 재생하려는 파일이 도중에(홈키를 눌러서 파일조작) 사라진경우 예외처리

+++ app/src/main/java/com/kp/appropriatebgm/DBController/DBManager.java
@@ -621,7 +621,8 @@ public class DBManager extends SQLiteOpenHelper {
         for(int i=0; i<bgmPath.length; i++){
             try {
                 file = new File(bgmPath[i]);
-                file.delete();
+                if (file.isFile())
+                    file.delete();
             } catch (Exception e){
                 Log.e("deleteBGMFile 파일삭제", e.toString());
             }
